Encuesta Longitudinal de Empresas

ELE data for the year 2022. Contains only a few variables.

Usage

ELE7

Format

dataframe with 6.592 rows y 13 columns

rol_ficticio

Company ID

cod_actividad

Economic activity

cod_tamano

Company size by sales

tramo

Inclusion range

fe_transversal

Cross-sectional weights

fe_longitudinal

Longitudinal weights

panel

Panel sample

estrato

Strata

pob

Finite population correction

VA_2022

Value added 2022, difference between production value and intermediate consumption

VA_2022f

VA_2022f is an adjusted version of VA_2022, where negative values are replaced with 0, while non-negative values remain unchanged.

EMP

Total personnel employed and hired by the company on a monthly basis

REMP_TOTAL

Total gross remuneration of personnel hired by the company

Source

https://www.ine.gob.cl/docs/default-source/encuesta-longitudinal-de-empresas/bbdd/ele-2022/base-de-datos-ele7.csv?sfvrsn=1504c58d_4&download=true

Examples

data(ELE7)
#> Warning: data set 'ELE7' not found