internal function to calculate proportion estimations

internal function to calculate proportion estimations

Usage

create_prop_internal(
  var,
  domains = NULL,
  subpop = NULL,
  disenio,
  ci = FALSE,
  deff = FALSE,
  ess = FALSE,
  ajuste_ene = FALSE,
  rel_error = FALSE,
  log_cv = FALSE,
  unweighted = FALSE,
  standard_eval = TRUE,
  rm.na = FALSE,
  env = parent.frame(),
  ci_logit = FALSE
)

Arguments

var

integer dummy variable within the dataframe

domains

domains to be estimated separated by the + character.

subpop

integer dummy variable to filter the dataframe

disenio

complex design created by survey package

ci

boolean indicating if the confidence intervals must be calculated

deff

boolean Design effect

ess

boolean Effective sample size

ajuste_ene

boolean indicating if an adjustment for the sampling-frame transition period must be used

rel_error

boolean Relative error

log_cv

boolean indicating if the log cv must be returned

unweighted

boolean Add non weighted count if it is required

standard_eval

boolean indicating if the function is inside another function, by default it is TRUE, avoid problems with lazy eval.

rm.na

boolean indicating if NA values must be removed

env

parent environment to get some variables

ci_logit

boolean indicating if interval confidence is logit

Value

dataframe that contains the inputs and all domains to be evaluated