Evaluate the quality of mean estimations

evaluate evauates the quality of mean estimation using the methodology created by INE Chile, which considers sample size, degrees of freedom and coeficient of variation.

Usage

evaluate(table, publish = FALSE, scheme = c("chile", "cepal"), ...)

Arguments

table

dataframe created by crear_insumos_media

publish

boolean indicating if the evaluation of the complete table must be added. If it is TRUE, the function adds a new column to the dataframe

scheme

string variable, default scheme is "chile" which refers to the evaluation protocol proposed by INE Chile. the alternative is "cepal" to use the CEPAL protocol

...

the list of cepal parameters. The complete list of parameters is

1. General Parameters

df degrees of freedom. default: 9

n sample size. default ine scheme is 60. default cepal scheme: 100

2. INE parameters

cv_lower_ine lower limit for cv. default: 0.15

cv_upper_ine upper limit for cv. default: 0.3

3. CEPAL parameters

cv_cepal limit for cv. default: 0.2

ess efective sample size. default: 140

unweighted unweighted count. default: 50

Value

dataframe with all the columns included in the input table, plus a new column containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable

Examples

dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
evaluate(create_mean("gastot_hd", domains = "zona+sexo", design = dc))
#>   zona sexo    stat       se  df     n         cv                 eval_n
#> 1    1    1 1378016 69421.07 127 12300 0.05037756 sufficient sample size
#> 2    2    1 1082227 36324.22 121 10393 0.03356434 sufficient sample size
#> 3    1    2 1325015 68510.58 127 13729 0.05170551 sufficient sample size
#> 4    2    2 1088658 35343.76 121 11886 0.03246545 sufficient sample size
#>         eval_df    eval_cv  label
#> 1 sufficient df cv <= 0.15 fiable
#> 2 sufficient df cv <= 0.15 fiable
#> 3 sufficient df cv <= 0.15 fiable
#> 4 sufficient df cv <= 0.15 fiable